rp/requirements.txt
# Requirements for Raspberry Pi
numpy>=1.21.0
librosa>=0.9.2
tensorflow-lite-runtime>=2.8.0
soundfile>=0.11.0
sounddevice>=0.4.4